// src/generated/client/src/models/MenuItemData.ts
export interface MenuItemData {
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_author?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_date?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_date_gmt?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_content?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_title?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_excerpt?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_status?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    comment_status?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    ping_status?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_password?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_name?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    to_ping?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_modified?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuItemData
     */
    post_modified_gmt?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_content_filtered?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuItemData
     */
    post_parent?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    guid?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    menu_order?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_type?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    post_mime_type?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    comment_count?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    filter?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuItemData
     */
    db_id?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    menu_item_parent?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    object_id?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    object?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    type_label?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    target?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    attr_title?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemData
     */
    xfn?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof MenuItemData
     */
    classes?: Array<string>;
    /**
     * 
     * @type {Array<MenuItemData>}
     * @memberof MenuItemData
     */
    child_items?: Array<MenuItemData>;
}